[es6-bookreader-core/Toolbar] create Toolbar.js
